Refactor sync module and introduce state management enhancements
- Moved sync-related functionality into dedicated modules, including `coordinator.rs`, `queue.rs`, `state.rs`, and `state_lock.rs`, improving organization and readability.
- Implemented a new `SyncCoordinator` struct to manage task state and synchronization processes more effectively.
- Added functions for task management, including `enqueue`, `changed_tasks`, and `merge_existing_task`, enhancing task tracking and state updates.
- Introduced state locking mechanisms to prevent concurrent access issues during sync operations.
- Updated the `PersistedSyncQueueState` structure to support better state persistence and task management.
- Enhanced error handling and logging for improved debugging and reliability in sync operations.
